+++ src/main/java/com/faforever/gw/security/User.java
@@ -1,32 +1,21 @@
 package com.faforever.gw.security;
 
 import com.faforever.gw.model.GwCharacter;
-import com.fasterxml.jackson.core.type.TypeReference;
-import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.Getter;
 import lombok.Setter;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.Authentication;
 import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.jwt.Jwt;
-import org.springframework.security.jwt.JwtHelper;
-import org.springframework.security.jwt.crypto.sign.MacSigner;
 
-import java.security.Principal;
-import java.util.ArrayList;
 import java.util.Collection;
-import java.util.List;
 
 @Getter
 @Setter
 public class User extends UsernamePasswordAuthenticationToken {
 
-  private final int id;
+    private final long id;
   private GwCharacter activeCharacter;
 
-  public User(int id, GwCharacter activeCharacter, Object principal, Object credentials, Collection<? extends GrantedAuthority> authorities) {
+    public User(long id, GwCharacter activeCharacter, Object principal, Object credentials, Collection<? extends GrantedAuthority> authorities) {
     super(principal, credentials, authorities);
     this.id = id;
     this.activeCharacter = activeCharacter;
